requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.0.7
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.13.1
flashlight-text==0.0.7
fonttools==4.55.7
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hypothesis==6.124.7
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.3
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=d23d300d559a45a5e3ede958dbbf2395231119c0b8cd97a1ea43480625894ff4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
networkx==3.2.1
numpy==1.24.4
openai==1.60.2
opencv-python==4.5.5.64
orjson==3.10.15
packaging==24.2
pandas==1.4.4
peft==0.13.2
pillow==10.2.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.3
safetensors==0.5.2
scipy==1.8.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==12.0
zipp==3.21.0
